4.21*: Blog List Expansion, step 9

+++ part4/blogs/backend/controllers/blogs.js
@@ -47,6 +47,14 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== decodedToken.id.toString()){
+    return response.status(401).json({
+      error: 'only the creator can delete the blog'
+    })
+
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
